feat(docker): build script for BO
BO is a customized LLVM which will be used to create smart contract programming languge on bitconch
chain.

BREAKING CHANGE: BO LLVM will help user to write smart contract. This repo provide a build script to
build the LLVM binaries. All changes to the LLVM should be done on the corresponding repos, as this
repo only provide building scripts.